+++ moneyplant-ui/src/app/features/dashboard/overall/overall.component.ts
@@ -1,12 +1,8 @@
 import { Component, OnInit, ChangeDetectorRef, ElementRef, ViewChild, ChangeDetectionStrategy, OnDestroy } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CardModule } from 'primeng/card';
 import { ButtonModule } from 'primeng/button';
-import { TagModule } from 'primeng/tag';
-import { DividerModule } from 'primeng/divider';
 import { MessageModule } from 'primeng/message';
-import { TooltipModule } from 'primeng/tooltip';
-import { GridsterConfig, DisplayGrid, GridType } from 'angular-gridster2';
+import { ScrollPanelModule } from 'primeng/scrollpanel';
 
 // Import echarts core module and components
 import * as echarts from 'echarts/core';
@@ -73,50 +69,10 @@ import('echarts-map-collection/custom/world.json').then((worldMapData) => {
   // Handle world map loading error silently
 });
 
-// Example of registering a custom Hong Kong map (if you have the GeoJSON data)
-// You can uncomment and modify this if you have Hong Kong GeoJSON data
-/*
-const hongKongGeoJson = {
-  "type": "FeatureCollection",
-  "features": [
-    // Your Hong Kong GeoJSON data here
-  ]
-};
-DensityMapBuilder.registerMap('HK', hongKongGeoJson);
-*/
-
 // Import dashboard modules and chart builders
 import { 
   IWidget,
   DashboardContainerComponent,
-  WidgetBuilder,
-  // Chart Builders
-  PieChartBuilder,
-  BarChartBuilder,
-  LineChartBuilder,
-  ScatterChartBuilder,
-  GaugeChartBuilder,
-  HeatmapChartBuilder,
-  AreaChartBuilder,
-  PolarChartBuilder,
-  StackedAreaChartBuilder,
-  TreemapChartBuilder,
-  SunburstChartBuilder,
-  SankeyChartBuilder,
-  // Data interfaces
-  PieChartData,
-  BarChartData,
-  LineChartData,
-  ScatterChartData,
-  GaugeChartData,
-  HeatmapChartData,
-  DensityMapData,
-  AreaChartData,
-  PolarChartData,
-  StackedAreaSeriesData,
-  TreemapData,
-  SunburstChartData,
-  SankeyChartData,
   // Fluent API
   StandardDashboardBuilder,
   DashboardConfig,
@@ -154,147 +110,26 @@ import {
   createBudgetAllocationSankeyWidget,
   createMinimalSankeyChartWidget,
   createFilterWidget,
-  // Data update functions
-  updateAssetAllocationData,
-  updateMonthlyIncomeExpensesData,
-  updatePortfolioPerformanceData,
-  updateRiskReturnData,
-  updateSavingsGoalData,
-  updateSpendingHeatmapData,
-  updateInvestmentDistributionData,
-  updateAreaChartData,
-  updatePolarChartData,
-  updateStackedAreaChartData,
-  updateTreemapChartData,
-  updateSunburstChartData,
-  updateSankeyChartData,
-  // Data fetching functions
-  getUpdatedAssetAllocationData,
-  getUpdatedMonthlyData,
-  getUpdatedPortfolioData,
-  getUpdatedRiskReturnData,
-  getUpdatedSavingsGoalData,
-  getUpdatedSpendingHeatmapData,
-  getUpdatedInvestmentDistributionData,
-  getUpdatedAreaChartData,
-  getUpdatedPolarChartData,
-  getUpdatedStackedAreaChartData,
-  getUpdatedTreemapChartData,
-  getUpdatedSunburstChartData,
-  getUpdatedSankeyChartData,
-  // Alternative data functions
-  getAlternativeAssetAllocationData,
-  getAlternativeMonthlyData,
-  getAlternativePortfolioData,
-  getAlternativeRiskReturnData,
-  getAlternativeSavingsGoalData,
-  getAlternativeSpendingHeatmapData,
-  getAlternativeInvestmentDistributionData,
-  getAlternativeAreaChartData,
-  getAlternativePolarChartData,
-  getAlternativeStackedAreaChartData,
-  getAlternativeTreemapChartData,
-  getAlternativeSunburstChartData,
-  getAlternativeSankeyChartData
+  // Dashboard data
+  DashboardDataRow,
+  INITIAL_DASHBOARD_DATA
 } from './widgets';
 
 // Import test filter widget directly
-import { createTestFilterWidget, updateTestFilterData } from './widgets/test-filter-widget';
+import { createTestFilterWidget } from './widgets/test-filter-widget';
 
 // Filter service
 import { FilterService } from '@dashboards/public-api';
 
-import { v4 as uuidv4 } from 'uuid';
-import { ScrollPanelModule } from 'primeng/scrollpanel';
-import { updatePieChartDataDirect } from './widgets/asset-allocation-widget';
 import { Subscription } from 'rxjs';
-
-// Shared data model for all widgets - Flat structure
-export interface DashboardDataRow {
-  id: string;
-  assetCategory: string;  // For Asset Allocation, Risk vs Return
-  month: string;          // For Monthly Income/Expenses, Portfolio Performance
-  market: string;         // For Investment Distribution map
-  totalValue: number;     // Primary value for most charts
-  riskValue?: number;     // For Risk vs Return scatter
-  returnValue?: number;   // For Risk vs Return scatter
-  description?: string;   // Additional context
-}
-
-// Initial dashboard data - Flat structure
-const INITIAL_DASHBOARD_DATA: DashboardDataRow[] = [
-  // Asset Allocation & Risk vs Return data (same categories for all charts)
-  { id: '1', assetCategory: 'Stocks', month: 'Jan', market: 'US', totalValue: 45, riskValue: 0.12, returnValue: 0.15 },
-  { id: '2', assetCategory: 'Bonds', month: 'Jan', market: 'US', totalValue: 25, riskValue: 0.05, returnValue: 0.08 },
-  { id: '3', assetCategory: 'Cash', month: 'Jan', market: 'US', totalValue: 15, riskValue: 0.03, returnValue: 0.05 },
-  { id: '4', assetCategory: 'Real Estate', month: 'Jan', market: 'US', totalValue: 10, riskValue: 0.08, returnValue: 0.10 },
-  { id: '5', assetCategory: 'Commodities', month: 'Jan', market: 'US', totalValue: 5, riskValue: 0.20, returnValue: 0.25 },
-  
-  // Monthly Income/Expenses data (using same categories)
-  { id: '6', assetCategory: 'Stocks', month: 'Feb', market: 'US', totalValue: 48, riskValue: 0.13, returnValue: 0.16 },
-  { id: '7', assetCategory: 'Bonds', month: 'Feb', market: 'US', totalValue: 26, riskValue: 0.06, returnValue: 0.09 },
-  { id: '8', assetCategory: 'Cash', month: 'Feb', market: 'US', totalValue: 16, riskValue: 0.04, returnValue: 0.06 },
-  { id: '9', assetCategory: 'Real Estate', month: 'Feb', market: 'US', totalValue: 11, riskValue: 0.09, returnValue: 0.11 },
-  { id: '10', assetCategory: 'Commodities', month: 'Feb', market: 'US', totalValue: 6, riskValue: 0.21, returnValue: 0.26 },
-  
-  // Portfolio Performance data (using same categories)
-  { id: '11', assetCategory: 'Stocks', month: 'Mar', market: 'US', totalValue: 50, riskValue: 0.14, returnValue: 0.17 },
-  { id: '12', assetCategory: 'Bonds', month: 'Mar', market: 'US', totalValue: 27, riskValue: 0.07, returnValue: 0.10 },
-  { id: '13', assetCategory: 'Cash', month: 'Mar', market: 'US', totalValue: 17, riskValue: 0.05, returnValue: 0.07 },
-  { id: '14', assetCategory: 'Real Estate', month: 'Mar', market: 'US', totalValue: 12, riskValue: 0.10, returnValue: 0.12 },
-  { id: '15', assetCategory: 'Commodities', month: 'Mar', market: 'US', totalValue: 7, riskValue: 0.22, returnValue: 0.27 },
-  
-  // Additional months for time series
-  { id: '16', assetCategory: 'Stocks', month: 'Apr', market: 'US', totalValue: 52, riskValue: 0.15, returnValue: 0.18 },
-  { id: '17', assetCategory: 'Bonds', month: 'Apr', market: 'US', totalValue: 28, riskValue: 0.08, returnValue: 0.11 },
-  { id: '18', assetCategory: 'Cash', month: 'Apr', market: 'US', totalValue: 18, riskValue: 0.06, returnValue: 0.08 },
-  { id: '19', assetCategory: 'Real Estate', month: 'Apr', market: 'US', totalValue: 13, riskValue: 0.11, returnValue: 0.13 },
-  { id: '20', assetCategory: 'Commodities', month: 'Apr', market: 'US', totalValue: 8, riskValue: 0.23, returnValue: 0.28 },
-  
-  { id: '21', assetCategory: 'Stocks', month: 'May', market: 'US', totalValue: 55, riskValue: 0.16, returnValue: 0.19 },
-  { id: '22', assetCategory: 'Bonds', month: 'May', market: 'US', totalValue: 29, riskValue: 0.09, returnValue: 0.12 },
-  { id: '23', assetCategory: 'Cash', month: 'May', market: 'US', totalValue: 19, riskValue: 0.07, returnValue: 0.09 },
-  { id: '24', assetCategory: 'Real Estate', month: 'May', market: 'US', totalValue: 14, riskValue: 0.12, returnValue: 0.14 },
-  { id: '25', assetCategory: 'Commodities', month: 'May', market: 'US', totalValue: 9, riskValue: 0.24, returnValue: 0.29 },
-  
-  { id: '26', assetCategory: 'Stocks', month: 'Jun', market: 'US', totalValue: 58, riskValue: 0.17, returnValue: 0.20 },
-  { id: '27', assetCategory: 'Bonds', month: 'Jun', market: 'US', totalValue: 30, riskValue: 0.10, returnValue: 0.13 },
-  { id: '28', assetCategory: 'Cash', month: 'Jun', market: 'US', totalValue: 20, riskValue: 0.08, returnValue: 0.10 },
-  { id: '29', assetCategory: 'Real Estate', month: 'Jun', market: 'US', totalValue: 15, riskValue: 0.13, returnValue: 0.15 },
-  { id: '30', assetCategory: 'Commodities', month: 'Jun', market: 'US', totalValue: 10, riskValue: 0.25, returnValue: 0.30 },
-  
-  // Test Filter data (using same categories)
-  { id: '31', assetCategory: 'Stocks', month: 'Jan', market: 'Europe', totalValue: 35, riskValue: 0.10, returnValue: 0.12 },
-  { id: '32', assetCategory: 'Bonds', month: 'Jan', market: 'Europe', totalValue: 20, riskValue: 0.04, returnValue: 0.06 },
-  { id: '33', assetCategory: 'Cash', month: 'Jan', market: 'Europe', totalValue: 12, riskValue: 0.02, returnValue: 0.04 },
-  { id: '34', assetCategory: 'Real Estate', month: 'Jan', market: 'Europe', totalValue: 8, riskValue: 0.06, returnValue: 0.08 },
-  { id: '35', assetCategory: 'Commodities', month: 'Jan', market: 'Europe', totalValue: 4, riskValue: 0.18, returnValue: 0.22 },
-  
-  // Additional market data for map charts
-  { id: '36', assetCategory: 'Stocks', month: 'Jan', market: 'Asia', totalValue: 40, riskValue: 0.15, returnValue: 0.18 },
-  { id: '37', assetCategory: 'Bonds', month: 'Jan', market: 'Asia', totalValue: 18, riskValue: 0.06, returnValue: 0.09 },
-  { id: '38', assetCategory: 'Cash', month: 'Jan', market: 'Asia', totalValue: 14, riskValue: 0.03, returnValue: 0.05 },
-  { id: '39', assetCategory: 'Real Estate', month: 'Jan', market: 'Asia', totalValue: 9, riskValue: 0.09, returnValue: 0.11 },
-  { id: '40', assetCategory: 'Commodities', month: 'Jan', market: 'Asia', totalValue: 6, riskValue: 0.19, returnValue: 0.23 },
-  
-  { id: '41', assetCategory: 'Stocks', month: 'Jan', market: 'Emerging', totalValue: 25, riskValue: 0.20, returnValue: 0.25 },
-  { id: '42', assetCategory: 'Bonds', month: 'Jan', market: 'Emerging', totalValue: 12, riskValue: 0.08, returnValue: 0.11 },
-  { id: '43', assetCategory: 'Cash', month: 'Jan', market: 'Emerging', totalValue: 10, riskValue: 0.05, returnValue: 0.07 },
-  { id: '44', assetCategory: 'Real Estate', month: 'Jan', market: 'Emerging', totalValue: 7, riskValue: 0.12, returnValue: 0.14 },
-  { id: '45', assetCategory: 'Commodities', month: 'Jan', market: 'Emerging', totalValue: 5, riskValue: 0.22, returnValue: 0.27 }
-];
 
 @Component({
   selector: 'app-overall',
   standalone: true,
   imports: [
     CommonModule, 
-    CardModule,
     ButtonModule,
-    TagModule,
-    DividerModule,
     MessageModule,
-    TooltipModule,
     ScrollPanelModule,
     // Dashboard components
     DashboardContainerComponent
@@ -852,60 +687,6 @@ export class OverallComponent implements OnInit, OnDestroy {
       // Handle Excel export error silently
     } finally {
       this.isExportingExcel = false;
-    }
-  }
-
-  /**
-   * Update a single widget with new data
-   */
-  public async updateWidget(widgetId: string, newData: any): Promise<void> {
-    const widget = this.dashboardConfig.widgets.find(w => w.id === widgetId);
-    if (!widget) {
-      return;
-    }
-
-    try {
-      // Apply filters to the new data
-      const filteredData = this.filterService.applyFiltersToData([newData], this.filterService.getFilterValues());
-      
-      if (filteredData.length > 0) {
-        // Update the widget with filtered data
-        if (widget.config.component === 'echart') {
-          // For chart widgets, update the series data
-          const chartOptions = widget.config.options as any;
-          if (chartOptions.series && chartOptions.series.length > 0) {
-            chartOptions.series[0].data = filteredData;
-          }
-        }
-      }
-    } catch (error) {
-      // Handle widget update error silently
-    }
-  }
-
-  /**
-   * Update multiple widgets with new data
-   */
-  public async updateMultipleWidgets(widgets: IWidget[], data: any[]): Promise<void> {
-    try {
-      widgets.forEach((widget, index) => {
-        if (data[index]) {
-          if (widget.config?.component === 'echart') {
-            // Update chart data based on chart type
-            const chartOptions = widget.config.options as any;
-            if (chartOptions && chartOptions.series) {
-              // Update series data
-              chartOptions.series.forEach((series: any, seriesIndex: number) => {
-                if (data[index][seriesIndex]) {
-                  series.data = data[index][seriesIndex];
-                }
-              });
-            }
-          }
-        }
-      });
-    } catch (error) {
-      // Handle multiple widget update error silently
     }
   }
